requirements.txt
rasterio
geopandas
numpy
xarray
scikit-learn
xgboost
shap
opencv-python
scipy
matplotlib
seaborn
folium
h5py
joblib
Pillow
fastapi
uvicorn
pyproj
fiona
shapely
scikit-image
pandas
tqdm
jupyter
notebook
streamlit
streamlit-folium
plotly
